refactor: remove web seed skills helper
