+++ turbo/apps/web/src/__tests__/db-test-seeders/seed-skill-values.ts
@@ -1,12 +1,9 @@
 import { resolveSkillRef } from "@vm0/core/github-url";
-import { SEED_SKILLS } from "@vm0/core/zero-seed-skills";
 import type { skills } from "@vm0/db/schema/skill";
-
-export { SEED_SKILLS };
 
 /**
  * Build skill insert values from a list of skill names.
- * Shared by dev-seed and test helpers to avoid duplicated URL/frontmatter construction.
+ * Shared by web test setup and DB-direct seeders.
  */
 export function buildSeedSkillValues(
   names: readonly string[],
